refactor: test 데이터를 위한 클래스들을 패키지 분리
