+++ src/test/java/nextstep/ladder/domain/tobe/fixture/RightConnectLineCreateStrategy.java
@@ -1,4 +1,7 @@
-package nextstep.ladder.domain.tobe;
+package nextstep.ladder.domain.tobe.fixture;
+
+import nextstep.ladder.domain.tobe.Direction;
+import nextstep.ladder.domain.tobe.LineCreateStrategy;
 
 public class RightConnectLineCreateStrategy implements LineCreateStrategy {
